+++ CHANGELOG.md
@@ -10,10 +10,12 @@ IMPROVEMENTS:
   - New callback for notifying message queue message send success (ACK from destination node received). Callback URL can be set using POST `/node/callback` with property `message_queue_send_success_url`.
   - Change NDID APIs (`/ndid/update_node`, `/ndid/enable_node`, and  `/ndid/disable_node`) HTTP success response status code from 200 to 204.
   - New API: GET `/idp/request_message_padded_hash` for getting `request_message_padded_hash` used for creating mode 2 and 3 request signature (signed with accessor private key without padding) on response.
+- Add accessor in request reference group validation on IdP responses.
 
 BUG FIXES:
 
 - Fix missing request status update callback (request closed, request timed out) on IdP side for identity related requests.
+- Fix invalid IdP response signature check on RP and IdP nodes when signature is cryptographically valid but signed with accessor that is not in request reference group.
 
 ## 2.0.1 (June 24, 2019)